+++ lib/pleroma/web/mastodon_api/controllers/app_controller.ex
@@ -25,8 +25,6 @@ defmodule Pleroma.Web.MastodonAPI.AppController do
 
   plug(Pleroma.Web.ApiSpec.CastAndValidate)
 
-  @local_mastodon_name "Mastodon-Local"
-
   defdelegate open_api_operation(action), to: Pleroma.Web.ApiSpec.AppOperation
 
   @doc "POST /api/v1/apps"
@@ -47,7 +45,6 @@ defmodule Pleroma.Web.MastodonAPI.AppController do
       |> Maps.put_if_present(:user_id, user_id)
 
     with cs <- App.register_changeset(%App{}, app_attrs),
-         false <- cs.changes[:client_name] == @local_mastodon_name,
          {:ok, app} <- Repo.insert(cs) do
       render(conn, "show.json", app: app)
     end