+++ priv/static/packs/features/search.js
@@ -1,2 +1,0 @@
-(window.webpackJsonp=window.webpackJsonp||[]).push([[46],{832:function(a,e,c){"use strict";c.r(e);var r=c(0),d=(c(3),c(781)),i=c(1178);e.default=function(){return Object(r.a)("div",{className:"column search-page"},void 0,Object(r.a)(d.a,{}),Object(r.a)("div",{className:"drawer__pager"},void 0,Object(r.a)("div",{className:"drawer__inner darker"},void 0,Object(r.a)(i.a,{}))))}}}]);
-//# sourceMappingURL=search.js.map